simplify loadfile process
SMOD.load_file returns the helper function, or nil and an error. This pattern can be passed straight into assert, which will throw an error if the file doesn't exist. This isn't a problem, since we don't want the game running with half a loaded mod anyways.
If load_file *does* load correctly, then helper will be truthy. assert returns the first value, if it's truthy, so we can then just call it immediately after, without needing to create excess variables.