Change logic to KISS
Using `next()` as a way to determine if a table is empty is an absolutely bonkers way to check that. It works, but it's not even remotely obvious at first (or second, or third) glance what it's meant to be doing.

`next()` is a function meant for iterating over a table. If you were creating your own iterator, then it makes sense to use next. But if you want to see if a sequentially indexed table has any items, then it's best to follow KISS.