use reroll_button
The `reroll_button` UI entity is assigned to a variable, then regot a second time. Before, it was accessed within assign_var, scoped out to the file, then accessed back within the new function, but the same problem still stood. It was simply obfuscated by the winding logic.

Now, we can just use the entity. A microopt, since we don't need to fetch twice (although, does the `remove` method remove the entity from the node it's in? Or from the game? Maybe it needs to be reaccessed after removal? Unclear. Testing is required)